update login, signup UI
