+++ app/src/main/res/drawable/bg_all_round_black.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners android:radius="50dp"/>
+    <stroke android:width="1dp" android:color="@color/black"/>
+</shape>
