+++ app/src/main/res/drawable/bg_all_round_square.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/white"/>
+    <corners android:radius="4dp"/>
+    <stroke android:width="1dp" android:color="@color/grey128"/>
+</shape>
