+++ app/src/main/res/drawable/gradient1.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient android:startColor="#845ECB"
+        android:endColor="#4472C4"
+        android:angle="-45"/>
+</shape>
