+++ app/src/main/res/drawable/gradient2.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient android:startColor="#CE9FFC"
+        android:endColor="#7367F0"
+        android:useLevel="false"
+        android:angle="-45"/>
+</shape>
