--- app/src/main/res/drawable/ic_intro.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="512dp"
-    android:height="512dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-  <path
-      android:pathData="M-0.002,244.54L339.998,-0L511.722,0.276L339.998,520L-0.002,520z"
-      android:strokeAlpha="0"
-      android:strokeLineJoin="miter"
-      android:strokeWidth="0"
-      android:fillColor="#797766"
-      android:fillType="nonZero"
-      android:strokeColor="#75F9AA"
-      android:strokeLineCap="butt"/>
-</vector>
