+++ app/src/main/res/layout/fragment_today.xml
@@ -152,7 +152,7 @@
                 android:layout_marginTop="10dp"
                 android:padding="30dp"
                 android:gravity="center"
-                android:text="가나다라마바"
+                android:text=""
                 android:fontFamily="@font/eastseadokdo_regular"
                 android:textColor="@color/black"
                 android:textSize="30sp"
