+++ app/src/main/res/values/colors.xml
@@ -25,6 +25,8 @@
     <color name="green">#4CAF50</color>
     <color name="skyblue">#00BCD4</color>
     <color name="navy">#3F51B5</color>
+    <color name="sisun_blue">#4472C4</color>
+    <color name="sisun_purple">#845ECB</color>
     <color name="champaigngold">#C4B6A6</color>
     <color name="champaigngold_bright">#D1C2B2</color>
     <color name="champaigngold_shade">#857B71</color>
